+++ backend/src/main/java/com/ssafy/dash/oauth/presentation/OAuthSessionController.java
@@ -8,9 +8,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ssafy.dash.oauth.application.security.CustomOAuth2User;
+import com.ssafy.dash.oauth.infrastructure.security.CustomOAuth2User;
 import com.ssafy.dash.user.application.dto.UserResponse;
 import com.ssafy.dash.user.application.UserService;
+import com.ssafy.dash.oauth.presentation.dto.response.OAuthSessionResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
